refactor: Member 도메인 리팩토링
- 원래 회원/비회원을 구분해서 저장하려고 했으나 비회원을 관리하기 힘들어서 그냥 스펙아웃 하기로 결정
